[ADDED] jupiter-engine maven dependency [FIXED] tests grouped
